+++ demos_fbs/csmc_cifar10.py
@@ -13,8 +13,8 @@ from fbs.data.images import normalise_rgb
 from fbs.sdes import make_linear_sde, make_linear_sde_law_loss, StationaryConstLinearSDE, \
     StationaryLinLinearSDE, StationaryExpLinearSDE, reverse_simulator
 from fbs.sdes.simulators import doob_bridge_simulator
-from fbs.filters.csmc.csmc import csmc_kernel
-from fbs.filters.csmc.resamplings import killing
+from fbs.samplers.csmc.csmc import csmc_kernel
+from fbs.samplers.csmc.resamplings import killing
 from fbs.nn.models import make_st_nn
 from fbs.nn.unet_cifar import UNet
 from fbs.nn.utils import make_optax_kernel
