+++ demos_fbs/csmc_pikachu_2d.py
@@ -14,8 +14,8 @@ import flax.linen as nn
 from fbs.nn import sinusoidal_embedding
 from fbs.nn.models import make_simple_st_nn
 from fbs.utils import discretise_lti_sde
-from fbs.filters.csmc.csmc import csmc_kernel
-from fbs.filters.csmc.resamplings import killing
+from fbs.samplers.csmc.csmc import csmc_kernel
+from fbs.samplers.csmc.resamplings import killing
 from functools import partial
 
 # General configs
