+++ demos_fbs/fbs_gauss_pmcmc.py
@@ -8,8 +8,8 @@ import matplotlib.pyplot as plt
 import numpy as np
 
 from fbs.utils import discretise_lti_sde
-from fbs.filters.smc import pmcmc_kernel
-from fbs.filters import stratified
+from fbs.samplers.smc import pmcmc_kernel
+from fbs.samplers import stratified
 from functools import partial
 
 # General configs
