+++ fbs/samplers/csmc/_csmc2.py
