+++ tests/test_cond_resamplings.py
@@ -5,7 +5,7 @@ import numpy as np
 import numpy.testing as npt
 import pytest
 
-from fbs.filters.csmc.resamplings import multinomial, systematic, killing
+from fbs.samplers.csmc.resamplings import multinomial, systematic, killing
 
 RESAMPLINGS = {"multinomial": multinomial, "systematic": systematic, "killing": killing}
